backend/requirements.txt
# FinanceClinics - Backend Dependencies
# Python 3.10+

# Flask Framework
Flask==2.3.3
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7

# Authentication
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2

# Security & Validation
Flask-WTF==1.2.1
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
bleach==6.1.0
python-slugify==8.0.1

# Email
Flask-Mail==0.9.1

# Environment
python-dotenv==1.0.0

# WSGI Server
gunicorn==21.2.0

# Utilities
Pillow==10.1.0
requests==2.31.0
pandas==2.2.3
openpyxl==3.1.2
python-docx==0.8.11
reportlab==4.0.0
pdfplumber==0.11.8

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
